Add completion suggestions for HLSL semantics.
