--- source/slang/slang-language-server-collect-member.h
@@ -1,26 +1,0 @@
-// slang-language-server-collect-member.h
-#pragma once
-
-#include "slang-ast-all.h"
-#include "slang-syntax.h"
-#include "slang-check-impl.h"
-
-namespace Slang
-{
-
-struct MemberCollectingContext
-{
-    ASTBuilder* astBuilder;
-    List<Decl*> members;
-    bool includeInstanceMembers = true;
-    SharedSemanticsContext semanticsContext;
-    MemberCollectingContext(Linkage* linkage, Module* module, DiagnosticSink* sink)
-        : semanticsContext(linkage, module, sink)
-    {}
-};
-
-void collectMembersInTypeDeclImpl(MemberCollectingContext* context, DeclRef<Decl> declRef);
-
-void collectMembersInType(MemberCollectingContext* context, Type* type);
-
-} // namespace Slang
